// src/enums/csvFieldsKeys.ts
export enum CsvFieldsKeysEnum {
  PurchaseOrderNumber = 'Purchase Order Number',
  BudgetCode = 'Budget Code',
  BudgetAmount = 'Budget Amount (SAR)',
  ProjectName = 'Project Name',
  InvoiceNumber = 'Invoice Num',
  InvoiceDate = 'Invoice Date',
  InvoiceAmount = 'Invoice Amount (SAR)',
  PaymentAmount = 'Payment Amount (SAR)',
}